Method tree builder for java

+++ src/main/kotlin/me/drbaxr/spektrum/flexible/adapters/java/model/external/MethodJava.kt
@@ -1,8 +1,8 @@
 package me.drbaxr.spektrum.flexible.adapters.java.model.external
 
 data class MethodJava(
-    val id: Int,
+    val id: Long,
     val signature: String,
-    val callers: Set<Int>,
-    val calledMethods: Set<Int>
+    val callers: Set<Long>,
+    val calledMethods: Set<Long>
 )
